Dev (#21)
* Updates to put the individual CCDs in DECam_CCD

* Modify SwarpSetup so -all does not have to be T whatever

* Addition of a new routine that allows one to make images
at an arbitray positions.

* Addition of a file containging lmc snrs, that can be used
to create images centred on the positions of SNRs

* Add a stand alone routine to produces image cutouts

* Directory related changes

* Variious updatesto handle new directory structure

* Small change

* SMC SNRs

* Update to SMC SNRs

* Fix problem on array shape in Cutout.py

* Cleanup of documentation and addtion of a
consistent set of switches

* Clean up and documentation

* Fix small error in log.py

* Better inut checking  of SwarpEval

* Add a new routine tha constructs region files
of all of the images in a directory

* Added a diagnostic file for plotting x-cuts
of a group of imagges at a specific RA and Dec.

* Updates to MakeXCut

* More fixes to MakeXcut

* To BackCalc.py add code to check how good the fits
are.

* Modified version of MakeXcut

* xMakeXcut functions, but it is not obvious
it is finding all of the images I hoped
it would find
	modified:   xMakeXcut.py

* Cleanup

* Updates to fix scaling and change location of plots

* Update to README

* Add a region file

* Fix to SwarpSetup (#19)

Previously with the -bsub option SwaprSetup pointed
to the files that were only had a mean background
per exposure subtracted.  With this change, it
creates input files that point to files in
DECam_PREP2 which is what was intended.